requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# ML/AI
torch
torchvision
transformers==4.35.0
scikit-learn
numpy==1.24.3
Pillow==10.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]
python-dotenv==1.0.0

# Utilities
redis==5.0.1
celery==5.3.4
boto3==1.34.0  # For AWS S3

# Monitoring
prometheus-client==0.19.0
python-json-logger==2.0.7

# Image processing
opencv-python==4.8.1
imageio==2.31.6